requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv

qdrant-client
openai
transformers
sentence-transformers
pymupdf
python-docx
pandas
